+++ src/pages/LoginPage/styled.js
@@ -1,0 +1,27 @@
+import styled from "styled-components";
+export const ContainerInput = styled.div`
+    display: grid;
+    justify-content: center;
+    align-items: center;
+    width: 100vw;
+    height: 100%;
+    form{
+        display: grid;
+        width: 90vw;
+        height: 30vh;
+        
+    }
+`
+
+export const ContairnerImg = styled.div`
+    display: grid;
+    justify-content: center;
+    align-items: center;
+    margin: 10px;
+    img{
+        display: grid;
+        justify-content: center;
+        align-items: center;
+    }
+
+`
